chore(layouts): move NavBar into its own folder
